Move interwiki link creation to test
While chasing down T277362 yesterday I was trying to debug this locally
and ended up with these changes that moves the interwiki link creation
from the extra-install scripts to the actual test in the hopes to
understand why things were failing. It didn't but i think this change is
preferable than having things hidden in extra-install.