Push to GHCR: continue-on-error
Set the non-main steps that push to github container
registry to continue-on-error. We always
retain an artefact with the tar.gz image this is a nice to
have rather than an essential so we can accept failure.

n.b. it seems that continue-on-error cannot be set in a composite
action so this was set in the workflow