Allow version/.env to specify extensions
This adds support for specifying which extensions should be installed
and enabled when building the wikibase bundle. This is a requirement in
order to be able to do future 1.35.N releases using this same release
pipeline.

This also adds the INSTALLED_EXTENSIONS env variable to the built docker
containers in order to keep track of which extensions was installed when
testing.