Run tests on example/
1. This adds the test-example make target that works similarly to test
but loads the images specified in the template.env in the tempfolder
instead.

2. This refactors run_selenium.sh so that DEFAULT_SUITE_CONFIG is
defined by the caller instead. This way it's easy to extend and makes it
more clear that test_suite.sh runs on the default docker-compose.yml and
any other caller would need to define their own.

3. Split out the hardcoded env variables defined by only being called by
test_suite.sh to default.env and example.env.

4. Refactor some of the code in test_stop.sh to in a similar way as in
point (2) so that all the make targets can use them. This also stops
using down in favor of kill as this is much faster.