show wikibase name in Quickstatements
until now, when creating a new batch in Quickstatements above the input field it said: "Create new command batch for" followed by an empty dropdown menu. Now the name of the wikibase instance is shown here. Though the user can not choose an alternative, but at least they know they are working on the correct wikibase instance. This is in particular important when one works on multiple different wikibase instances.