Github Actions, on PR and push (+selectivity)
This makes CI run on PRs as well, rather than only on the
underlying branches.
This will lead to the PR jobs actually running with the base of
the PR merged into the head.

Also be more selective about when to run some jobs, so for example
we can now make docs changes, without running the WHOLE build and
test pipeline.