docs: publishing: Update example images before tagging
I ran into this issue when doing the wmde.5 release, where
I followed the docs, tagged the image, and then when trying
to update links to the example I realized that I could not
link to the tag (what was done previously).

Looking at the git history this is also not the first time
this has happened, and folks have deleted tags on the repo
so they could re tag after then updating the example.

So document this step in the order of releasing.