Add CHANGE.md showing key facts about the included releases
This is a document targeting developers that have to work
with this repository to give them an overview of what has
happened when, and what has been changing and why.

I was not able to get such an overview by looking at the
existing versioned .env files, or build output (which is
mostly gone and not linked to), or other release notes, such
as those for Wikibase.git as every is very dispersed and not linked.

I hope that this doc might serve as a good overview, but also
is NOT targeting users, though some may find a high level
overview such as this useful.

We also want a location to write about changes to the docker
images for example, perhaps this is that place? perhaps it is
elsewhere?