Docs: Rewrite publishing.md as run through again
When initially running through a publish I had to get some
outside help (tobias and tom).
I managed to get through the second run by following the
docs that are in this commit.

Some things were in the wrong order before, for example you must
set your github api token before being able to download artifacts,
but this env var wasn't detailed until half way down the file.

Also seemingly some developers would not be able to run the tar
publishing setup using the documented methods, for example if not
having their ssh keys on disk, or using a yubi key.
Thus I publish an alternative path for this step in the commit.