feat: extend gzip compression rate usage
The release pipeline has a setting called GZIP_COMPRESSION_RATE that can
be set in local.env. Using this setting gzip invocations are configured.
The user can choose the compression rate/speed tradeoff. 1 being fast
and weak compression, 9 being slow and strong compression. This change
extends the usage of this configuration value to the remaining gzip
invocations as well as fixes the handing over of the actual value to
gzip by removing the extra 'f' (not required or supported according to
man gzip(1)).