fix: remove world write bit from /var/www/html
Without this change, /var/www/html would have drwxrwxrwt permissions,
that is in particular world writable with sticky bit set. This in turn
only allows the owner of a link in that directory to follow that link.
Wikibase uses a symlink "w" in the web root to link back to the web root
as a script path. In order to have apache be able to follow this link,
either the sticky bit needs to be removed, the link has to be owned by
www-data or the world writable flag needs to be removed. This patch does
the latter, as having 777 on a web root is considered bad practice
anyway.