fix: clean log as non root user
test/docker-compose.yml is mounting test/log/$SUITE/wikibase for the
webserver to write its logs into it as a volume. As the wikibase sub
directory did not exist, docker created it with root ownership. So the
system user actually running docker compose was not allowed to delete the
directory on the next run to clean logs. This change creates the
directory as the system user before docker compose is run, so docker
compose finds an existing directory and does not create it. As the
system user created it, the system user can also delete it afterwards.