feat(wdqs)!: allow setting of concept URI
Allow setting the concept URI of entities in WDQS explicitly and
distinct from the hostname used to contact wikibase.

This allows the concept URI to be a public URI while the wikibase
hostname is the docker network internal hostname used to poll for
updates.

BREAKING CHANGE: env var WIKIBASE_CONCEPT_URI now mandatory to run image