fix: drop privileges during docker build (#773)
Until now, the WDQS image started as root. In the entrypoint priveleges
were dropped and we switched to the "blazegraph" user. With this patch,
we switch to the "blazegraph" user already during build. So the
container always uses this user and does not start as root anymore.

This fixes some issues when running an interactive bash inside the
container and simplifies the updater reset workaround.

https://phabricator.wikimedia.org/T362963