Use SoftwareInfo mediawiki hook in suite and deploy
This will allow identification of both suite and deploy
seperatly in future released, via tooling such as
the wikibase-metatdata scraper, and other scrapers
such as for wikibase.world.

This could be implemented as a configurable extension that
would then be included as part of suite and deploy, however
that seemed a bit over the top when I started, given
the implementation here is small.

Bug: T411587